feat create components to explorer
