+++ frontend/src/constants/categoriaMock.js
@@ -1,0 +1,1 @@
+export const categorias = ["Todas", "Livros", "Brinquedos", "Roupas"];
