+++ frontend/src/constants/itensMock.js
@@ -1,0 +1,50 @@
+export const itensDisponiveis = [
+    {
+        id: 1,
+        titulo: "Livros de Romance",
+        descricao: "Coleção de 15 livros de romance em ótimo estado.",
+        categoria: "Livros",
+        imagem: "https://images.unsplash.com/photo-1544716278-ca5e3f4abd8c?q=80&w=2574&auto=format&fit=crop",
+        usuario: "Maria Silva",
+        bairro: "Centro",
+        dataPublicacao: "2 dias atrás",
+        rating: 4.5,
+        likes: 12,
+    },
+    {
+        id: 2,
+        titulo: "Bicicleta Infantil",
+        descricao: "Bicicleta aro 16, cor azul, pouco usada.",
+        categoria: "Brinquedos",
+        imagem: "https://images.unsplash.com/photo-1723322699174-ea81b9b2d1c5?q=80&w=1032&auto=format&fit=crop&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+        usuario: "João Santos",
+        bairro: "Vila Nova",
+        dataPublicacao: "1 dia atrás",
+        rating: 4.9,
+        likes: 8,
+    },
+    {
+        id: 3,
+        titulo: "Roupas Femininas Tam M",
+        descricao: "Lote com 10 peças de roupas femininas tamanho M.",
+        categoria: "Roupas",
+        imagem: "https://images.unsplash.com/photo-1594633312681-425c7b97ccd1?q=80&w=2574&auto=format&fit=crop",
+        usuario: "Ana Costa",
+        bairro: "Jardim das Flores",
+        dataPublicacao: "3 dias atrás",
+        rating: 4.7,
+        likes: 21,
+    },
+    {
+        id: 4,
+        titulo: "Box Sherlock Holmes",
+        descricao: "Edição de luxo com todos os contos e romances.",
+        categoria: "Livros",
+        imagem: "https://images.unsplash.com/photo-1749849689603-7fab16de47bb?q=80&w=736&auto=format&fit=crop&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+        usuario: "Pedro Martins",
+        bairro: "Centro",
+        dataPublicacao: "5 dias atrás",
+        rating: 4.8,
+        likes: 15,
+    },
+];
